refactor(canvas): extract skia_canvas_gradients.cpp (R2-3 second slice)
Phase 4 R2-3 follow-up. Pull the 194-line gradients + patterns paint
slice out of skia_canvas.cpp into a sibling TU. The slice covers:

- fill gradients (linear, radial, conic, two-circles)
- stroke gradients (linear, radial, conic, two-circles)
- fill + stroke image patterns (with repeat-mode + matrix)
- colors_to_skia private helper (used only by this slice)

- skia_canvas.cpp: 3,943 → 3,749 lines (-194 net).
- skia_canvas_gradients.cpp: 232 lines.
- Registered in core/canvas/CMakeLists.txt under PULP_HAS_SKIA.

Tests:
- pulp-canvas builds clean on macOS arm64 (Debug, GPU + Skia).
- pulp-test-canvas: 1,550 assertions in 63 cases — pass.
- pulp-test-canvas-emoji: 46 assertions in 13 cases — pass.

R2-3 progress: mask (first cut, prior PR) + gradients (this cut).
Remaining slices: box-shadow, opacity layers, GPU SDF, SkSL shaders,
plus the cg_canvas.mm mirror split.
