+++ core/canvas/CMakeLists.txt
@@ -79,6 +79,7 @@ if(PULP_HAS_SKIA)
     target_sources(pulp-canvas PRIVATE
         src/skia_canvas.cpp
         src/skia_canvas_mask.cpp
+        src/skia_canvas_gradients.cpp
         src/bundled_fonts.cpp
         # TextFontContext glues emoji_segmenter into the SkShaper +
         # SkParagraph paths and owns the per-context FontCollection used
